Move 3W region dropdown to its view
