--- src/components/domain/Region3WDropdown/styles.module.css
@@ -1,14 +1,0 @@
-.region-dropdown {
-    &.active {
-        text-decoration: underline;
-        color: var(--go-ui-color-primary-red);
-    }
-}
-
-.dropdown {
-    .menu-item {
-        &.active {
-            color: var(--go-ui-color-primary-red);
-        }
-    }
-}
