+++ src/components/domain/RiskImminentEvents/styles.module.css
@@ -22,7 +22,7 @@
                 }
 
                 .wildfire-icon {
-                    width: 60%;
+                    height: 0.8rem;
                 }
             }
         }
