+++ src/views/GlobalThreeW/RegionDropdown/i18n.json
@@ -1,5 +1,5 @@
 {
-    "namespace": "region3WDropdown",
+    "namespace": "globalThreeW",
     "strings": {
         "menu3WRegions":"Explore Regional 3W",
         "menuRegionsTooltip":"View 3W Regions",
