+++ src/views/GlobalThreeW/index.tsx
@@ -12,12 +12,12 @@ import BlockLoading from '#components/BlockLoading';
 import KeyFigure from '#components/KeyFigure';
 import Container from '#components/Container';
 import BarChart from '#components/BarChart';
-import Region3WDropdown from '#components/domain/Region3WDropdown';
 import PieChart from '#components/PieChart';
-import { useRequest } from '#utils/restRequest';
 import useTranslation from '#hooks/useTranslation';
+import { useRequest } from '#utils/restRequest';
 import { resolveToComponent } from '#utils/translation';
 
+import RegionDropdown from './RegionDropdown';
 import Filter, { type FilterValue } from './Filters';
 
 import {
@@ -31,6 +31,7 @@ import Map from './Map';
 import i18n from './i18n.json';
 import styles from './styles.module.css';
 
+// FIXME: use predefined colors
 const PIE_COLORS = ['#f64752', '#fa999f', '#fdd6d9'];
 
 function projectPerSecondarySectorsLabelSelector(
@@ -215,7 +216,7 @@ export function Component() {
                             />
                         )}
                         actions={(
-                            <Region3WDropdown />
+                            <RegionDropdown />
                         )}
                     >
                         <Map projectList={nsProjectsResponse} />
